requirements.txt
alpaca-trade-api==0.51.0
asgiref==3.3.1
astroid==2.2.5
attrs==21.4.0
beautifulsoup4==4.11.1
billboard.py==7.0.0
certifi==2020.11.8
chardet==3.0.4
click==8.0.3
dataclasses==0.8
Django==3.1.6
Flask==2.0.2
idna==2.10
importlib-metadata==4.8.3
iniconfig==1.1.1
isort==4.3.18
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.3.1
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.19.4
packaging==21.3
pandas==1.1.4
pluggy==1.0.0
py==1.11.0
pylint==2.3.1
pymongo==3.8.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.12.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
ta==0.7.0
tomli==1.2.3
typed-ast==1.3.5
typing-extensions==4.1.0
urllib3==1.25.11
websocket-client==0.57.0
websockets==8.1
Werkzeug==2.0.3
wrapt==1.11.1
zipp==3.6.0
